requirements.txt
# Required packages for Agentic AI with MCP
chainlit>=0.7.0
openai>=1.3.0
fastmcp>=0.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.0
typing-extensions>=4.7.0
pymongo>=4.6.0
uuid>=1.30
pytest>=7.4.0
loguru>=0.7.0
